+++ packages/client/CHANGELOG.md
@@ -5,6 +5,11 @@
 ### Added
 
 - `getMetadata$` and `getMetadata` client APIs to facilitate the retrieval of the most modern stable version of the metadata.
+- `BlockInfo` now exposes `hasNewRuntime` property.
+
+### Fixed
+
+- `blocks$` completes when there is no block continuity.
 
 ## 1.16.4 - 2025-09-01